refactored PaymentTransaction - tested - OK

+++ backend/src/main/java/com/pymntprocessing/pymntprocessing/service/PaymentTransactionImpl.java
@@ -1,5 +1,9 @@
 package com.pymntprocessing.pymntprocessing.service;
 
+import com.pymntprocessing.pymntprocessing.constant.db.TransactionTypeValue;
+import com.pymntprocessing.pymntprocessing.exception.InvalidDataProvidedException;
+import com.pymntprocessing.pymntprocessing.exception.PaymentTransactionNotFoundException;
+import com.pymntprocessing.pymntprocessing.model.entity.Vendor;
 import com.pymntprocessing.pymntprocessing.model.mapper.PaymentTransactionMapper;
 import com.pymntprocessing.pymntprocessing.model.dto.PaymentTransactionDTO;
 import com.pymntprocessing.pymntprocessing.model.entity.PaymentTransaction;
@@ -11,6 +15,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -18,22 +23,24 @@ public class PaymentTransactionImpl implements PaymentTransactionService{
 
     private final PaymentTransactionRepository paymentTransactionRepository;
     private final TransactionTypeRepository transactionTypeRepository;
+    private final VendorService vendorService;
 
     private final ModelMapper modelMapper;
 
     private final PaymentTransactionMapper paymentTransactionMapper;
 
     @Autowired
-    public PaymentTransactionImpl(PaymentTransactionRepository paymentTransactionRepository, TransactionTypeRepository transactionTypeRepository, ModelMapper modelMapper, PaymentTransactionMapper paymentTransactionMapper) {
+    public PaymentTransactionImpl(PaymentTransactionRepository paymentTransactionRepository, TransactionTypeRepository transactionTypeRepository, VendorService vendorService, ModelMapper modelMapper, PaymentTransactionMapper paymentTransactionMapper) {
         this.paymentTransactionRepository = paymentTransactionRepository;
         this.transactionTypeRepository = transactionTypeRepository;
+        this.vendorService = vendorService;
         this.modelMapper = modelMapper;
         this.paymentTransactionMapper = paymentTransactionMapper;
     }
 
     @Override
     public PaymentTransactionDTO getPaymentTransactionById(Long id) {
-        return this.paymentTransactionRepository.findById(id).map(this.paymentTransactionMapper::convertToDTO).orElse(null);
+        return this.paymentTransactionRepository.findById(id).map(this.paymentTransactionMapper::convertToDTO).orElseThrow(PaymentTransactionNotFoundException::new);
     }
 
     @Override
@@ -48,6 +55,32 @@ public class PaymentTransactionImpl implements PaymentTransactionService{
 
     @Override
     public PaymentTransactionDTO createPaymentTransaction(PaymentTransactionDTO paymentTransactionDTO) {
+        String errorMessage = "";
+
+        /**
+         * check if it's a valid vendor
+         */
+
+        Vendor existingVendor = this.vendorService.getVendorById(paymentTransactionDTO.getVendor().getId());
+
+        if (existingVendor == null) {
+            errorMessage = "Invalid vendor provided!";
+            throw new InvalidDataProvidedException(errorMessage);
+        }
+
+
+        /**
+         * check if it's a valid transaction type
+         */
+        TransactionType transactionType = paymentTransactionDTO.getTransactionType();
+        TransactionType existingTransactionType = this.getTransactionTypeById(paymentTransactionDTO.getTransactionType().getId());
+        if (existingTransactionType == null
+                || (!Objects.equals(transactionType.getName(), TransactionTypeValue.DEBIT.toString())
+                && !Objects.equals(transactionType.getName(), TransactionTypeValue.CREDIT.toString()))) {
+            errorMessage = "Invalid transaction type provided!";
+            throw new InvalidDataProvidedException(errorMessage);
+        }
+
         return paymentTransactionMapper.convertToDTO(this.paymentTransactionRepository.save(this.paymentTransactionMapper.convertToEntity(paymentTransactionDTO)));
     }
 
@@ -61,6 +94,32 @@ public class PaymentTransactionImpl implements PaymentTransactionService{
     @Override
     public PaymentTransactionDTO updatePaymentTransaction(Long id, PaymentTransactionDTO paymentTransactionDTO) {
 
+        String errorMessage = "";
+
+        /**
+         * check if it's a valid vendor
+         */
+
+        Vendor existingVendor = this.vendorService.getVendorById(paymentTransactionDTO.getVendor().getId());
+
+        if (existingVendor == null) {
+            errorMessage = "Invalid vendor provided!";
+            throw new InvalidDataProvidedException(errorMessage);
+        }
+
+
+        /**
+         * check if it's a valid transaction type
+         */
+        TransactionType transactionType = paymentTransactionDTO.getTransactionType();
+        TransactionType existingTransactionType = this.getTransactionTypeById(paymentTransactionDTO.getTransactionType().getId());
+        if (existingTransactionType == null
+                || (!Objects.equals(transactionType.getName(), TransactionTypeValue.DEBIT.toString())
+                && !Objects.equals(transactionType.getName(), TransactionTypeValue.CREDIT.toString()))) {
+            errorMessage = "Invalid transaction type provided!";
+            throw new InvalidDataProvidedException(errorMessage);
+        }
+
         Optional<PaymentTransaction> existingPaymentTransaction = this.paymentTransactionRepository.findById(id);
         if (existingPaymentTransaction.isPresent()) {
             return this.paymentTransactionMapper.convertToDTO(this.paymentTransactionRepository.save(this.paymentTransactionMapper.convertToEntity(paymentTransactionDTO)));
@@ -70,6 +129,11 @@ public class PaymentTransactionImpl implements PaymentTransactionService{
 
     @Override
     public void deletePaymentTransaction(Long id) {
+        PaymentTransactionDTO paymentTransactionDTO = this.getPaymentTransactionById(id);
+
+        if (paymentTransactionDTO == null) {
+            throw new InvalidDataProvidedException("Invalid id provided");
+        }
         this.paymentTransactionRepository.deleteById(id);
     }
 }
